DPA-89:  VeraPDF reenabled.  Process URL inputstream added.
There is a well-known optimization for the SB pillar that you can rewrite the
external URL to point to a file in a mounted directory.  We cannot use that for
the vagrant image as TRA has not been able to find a way for the host to
reliably access files inside the guest.

Hence the functionality has changed.  If the URL has the configured prefix
it is manipulated into a filename without being resolved, as before.  If not
the URL is used directly to obtain the input stream which should be a reasonable
fallback for a bitrepository client.  For the vagrant image initial testing
has shown that the trivial bitrepository web server crashes and must be restarted.

This has been filed as DPA-95.