DPA-35: Launchers split: Localhost+Achernar.  ConfigurationMap defaults from external file.
The IDE launchers were originally designed for just being used against the servers
running locally in a vagrant image.  Now we need to work against remote servers so
a split was prudent for both code and property files.
Also the configuration files now need to contain actual real user names and passwords,
so the current mechanism which allows for loading a single property file with extra
key-value pairs on the command line was not _quite_ flexible enough.

First a solution adding an include-mechanism to the standard Java property file was
considered but this got too cumbersome and did not easily allow for variable expansion.

The ability to as the first read in an external property file (with an arbitrary name set
with a system property) was decided to be the best solution for now with the least source
code changes.