DPA-35: Changed vagrant base image to bento/ubuntu-14.04 and added a 1 TB disk currently next to Vagrantfile.
Apparently the default is to have a fully qualified path to any diskimage file, and it was not
immediately evident how to locate the virtual machine used by vagrant to place the diskimage file there.
Therefore for now the file is just placed in the root of the sources, and ignored by git.
A question on how to solve this has been raised on https://stackoverflow.com/q/44115603/53897