DPA-35: Worked around appassembler limitation.
It turned out that appassembler plugin assumes that the "bin" folder is
located right under $BASEDIR (instead of in services/bin) in order for
the very simple REPO calculation to end up the right place.  The fix
was to tell appassembler to work in "appassembler/services" instead and
then have paths directly under that.