Reworking dashboard so listState.jsp now works correctly for what we have. CDI may not be possible.
The four entries on the listState.jsp page are now servlettified with common <%@include ... %>
presentation logic.  It is now possible to see present events, delete them and add a manual
control event.

It was found that CDI gave an obscure message (about nothing useful in request scope) which
was not easy to fix, so the current servlets for now just invoke the deprecated configuration.