After careful consideration moved CDI inside dpa-dashboard.
CDI is needed for dependency injection, and was used with Glassfish 4
which we do not currently use in production.  Hence a decision had to be made
on how to make things work.  Manual configuration, migrate to a CDI capable
container (Glassfish/TomEE) or add the necessary (but supported) glue to the
war file to make it work in plain Tomcat.

After careful consideration, the "make it work in plain Tomcat" approach was
chosen, and the result validated with Tomcat 8.5.16

Instructions found at http://balusc.omnifaces.org/2013/10/how-to-install-cdi-in-tomcat.html
(apparently a jandex dependency is now also needed).