Reenabled verapdf caching in DOMS to work around the file handling leak in VeraPDF.
Each invocation of VeraPDF saves the result in a VERAPDF data stream on the page.
After a crash the next run will use the cached output on the pages processed without
invoking VeraPDF so the next 4096 pages can be processed until eventually all have
been processed.  This is a work around unfortunately :-/