checksums-for-delivery-in-doms.sh generates md5sum from DOMS.
We need a simple ingest integrity checker, and the simplest
is to traverse DOMS for a given delivery round trip and
based on what is at the expected places construct a md5sum
file which can immediately be applied to the original
delivery tree on disk, or compared with "checksums.txt" (after
removing carriage returns).  Currently the Bitrepository
URL's for PDF files are expected to be valid (the URL
rewrite to disk mount point shortcut is not yet implemented).