HOTFIX:  Added BITREPOSITORY_INGESTER_MAX_RETRIES = "bitrepository.ingester.maxPutRetries"
It has turned out the the bit repository loses messages causing files not to be
reported as ingested which in turn mean that the ingest fails.  The official workaround
is to retry up to X times before reporting an error.    The actual number is passed
in as a configuration parameter "bitrepository.ingester.maxPutRetries" (using the same name)
which must be a valid integer.