DPA-124:  Move VeraPDF invocation to a separate process.
We have found during the project that VeraPDF have had leaks causing
the JVM to run out of resources and needing a restart.  Most
of these have been fixed, but to make the autonomous components
invoking VeraPDF more robust, the functionality must be
moved to a separate JVM and invoked as a simple web service.
This branch implements that.