Now checks events on SBOI matches to ensure that the items are as expected.
When adapting the newspaper autonomous components the SBOI lookup did not
incorporate the "double check that events are as expected" part so
items could be processed multiple times if the SBOI index was not updated
with the latest events.  This code has now been inlined so we can see in the log
if an entry was considered but skipped.