version3-release: ingest verification added.
Ingests are now verified against original deliveries by
traversing DOMS for the delivery and regenerating a
`md5sum` line for each article xml, page xml and page pdf.

If the lines are the same after sorting the ingest is
considered successful as we have what was intended.

The workflow in the configuration files
has been reworked so that the event "Checksumfile_checked"
is the trigger point for further processing.